better file names (#463)
